+++ packages/orq-rc/src/models/components/partdoneevent.ts
@@ -79,7 +79,7 @@ export const PartReasoningPart$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("reasoning_01kevdrr67yzq21s74r891v807"),
+  _id: z.string().default("reasoning_01keveeq318s7h31wa3yyasv1h"),
   metadata: z.record(z.any()).optional(),
   kind: PartKind$inboundSchema,
   reasoning: z.string(),
@@ -104,7 +104,7 @@ export const PartReasoningPart$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   PartReasoningPart
 > = z.object({
-  id: z.string().default("reasoning_01kevdrr67yzq21s74r891v807"),
+  id: z.string().default("reasoning_01keveeq318s7h31wa3yyasv1h"),
   metadata: z.record(z.any()).optional(),
   kind: PartKind$outboundSchema,
   reasoning: z.string(),