+++ packages/orq-rc/src/models/operations/createconversation.ts
@@ -270,7 +270,7 @@ export const CreateConversationResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("conv_01kevdrredpyp5vxpe13yymphd"),
+  _id: z.string().default("conv_01keveeqc66by3b4g2n3cs72hj"),
   kind: CreateConversationKind$inboundSchema,
   displayName: z.string(),
   createdAt: z.number(),
@@ -302,7 +302,7 @@ export const CreateConversationResponseBody$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   CreateConversationResponseBody
 > = z.object({
-  id: z.string().default("conv_01kevdrredpyp5vxpe13yymphd"),
+  id: z.string().default("conv_01keveeqc66by3b4g2n3cs72hj"),
   kind: CreateConversationKind$outboundSchema,
   displayName: z.string(),
   createdAt: z.number(),