+++ packages/orq-rc/src/models/operations/filelist.ts
@@ -157,7 +157,7 @@ export const FileListData$inboundSchema: z.ZodType<
   file_name: z.string(),
   workspace_id: z.string(),
   created: z.string().datetime({ offset: true }).default(
-    "2026-01-13T10:17:26.970Z",
+    "2026-01-13T10:29:23.759Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {
@@ -190,7 +190,7 @@ export const FileListData$outboundSchema: z.ZodType<
   bytes: z.number(),
   fileName: z.string(),
   workspaceId: z.string(),
-  created: z.date().default(() => new Date("2026-01-13T10:17:26.970Z"))
+  created: z.date().default(() => new Date("2026-01-13T10:29:23.759Z"))
     .transform(v => v.toISOString()),
 }).transform((v) => {
   return remap$(v, {